va tomando forma ya

+++ src/main/java/com/tuvieja/cart/service/CartService.java
@@ -4,39 +4,63 @@ import java.util.Collection;
 
 import javax.annotation.Resource;
 
+import org.joda.time.DateTime;
+
 import com.tuvieja.cart.dao.CartDao;
 import com.tuvieja.cart.dto.Cart;
-import com.tuvieja.cart.dto.Product;
+
+/*
+ * VERIFICACIONES:
+ * - Ver que existan los registros a modificar
+ * - Ver que existan las referencias (userId, productId, etc), llamar a los DAOs correspondientes 
+ * 
+ */
 
 public class CartService {
 	private @Resource CartDao cd;
 
-	public Cart fetchOne (String cartId){
-		return cd.fetchOne (cartId);
+	public Cart fetchOne(String cartId) {
+		return cd.fetchOne(cartId);
 	}
-	
-	public Collection<Cart> fetchAll (){
-		return cd.fetchAll ();
+
+	public Collection<Cart> fetchAll() {
+		return cd.fetchAll();
 	}
-	
-	public void createCart (Cart cart){
-		cd.createCart (cart);
+
+	public void createCart(Cart cart) {
+		// verificar que exista el usuario, llamar al UserDao
+		cart.setTimeStamp(DateTime.now());
+		cart.setCartStatus("cart created");
+		cd.createCart(cart);
 	}
-	
-	public void editCart (String cartId, Cart cart){
-		cd.editCart (cartId, cart);
-	}
-	
-	public void deleteCart (String cartId){
-		cd.deleteCart (cartId);
-	}
-	
-	public float getDiscount (String cartId) {
-		Cart cart = fetchOne (cartId);
-		float discount = 0f;
-		for (Product p : cart.getCartProducts()){
-			discount += Math.random() * p.getPrice();
+
+	public void buyCart(String cartId, Cart cart) {
+		// verificar que exista el carrito
+		// verificar que solo cambie el status (que lo hacemos acá)
+		if (cd.exists(cartId)) {
+			cart.setCartStatus("cart boght");
+			cd.editCart(cartId, cart);
 		}
-		return discount;
 	}
+
+	public void editCart(String cartId, Cart cart) {
+		// verificar que exista el carrito
+		// no permitir que cambie el usuario,
+		cd.editCart(cartId, cart);
+	}
+
+	public void deleteCart(String cartId) {
+		if (cd.exists(cartId)) {
+			cd.deleteCart(cartId);
+		}
+	}
+
+	// public float getDiscount (String cartId) {
+	// Cart cart = fetchOne (cartId);
+	// float discount = 0f;
+	// for (Product p : cart.getCartProducts()){
+	// discount += Math.random() * p.getPrice();
+	// }
+	// return discount;
+	// }
 }
